product-service/requirements.txt
PyJWT==2.1.0
requests==2.26.0
pyjwt[crypto]
cryptography